voice_control_gateway/requirements.txt
vosk==0.3.45
sounddevice==0.4.7
PyYAML==6.0.2
requests==2.32.3
